# Optimized std.datetime.date.Date.fromISOString