@JackStouffer 2017-05-20
@aG0aep6G Fixed the issue and managed to squeeze more performance out of it. Updated original comment