@JackStouffer 2017-05-20
@jmdavis I already fixed `to` to not auto decode. See https://github.com/dlang/phobos/pull/5014

>Try byCodeUnit, but I don't know if the bug preventing std.conv working with byCodeUnit was fixed yet

I also fixed that, see https://github.com/dlang/phobos/pull/4390