@jmdavis 2017-05-20
> I also fixed that, see #4390

Yay! Thanks.