@JackStouffer 2017-06-01
Ping @jmdavis, since you approved the other PR can you take a look at this?